refactor(consensus): create shared test utils and deduplicate code
