+++ consensus/istanbul/backend.go
@@ -37,13 +37,9 @@ type DecryptFn func(accounts.Account, []byte, []byte, []byte) ([]byte, error)
 // backing account.
 type SignerFn func(accounts.Account, string, []byte) ([]byte, error)
 
-// BLSSignerFn is a signer callback function to request a hash to be signed by a
-// backing account using BLS.
-type BLSSignerFn func(accounts.Account, []byte) (blscrypto.SerializedSignature, error)
-
-// BLSMessageSignerFn is a signer callback function to request a raw message to
-// be signed by a backing account.
-type BLSMessageSignerFn func(accounts.Account, []byte, []byte) (blscrypto.SerializedSignature, error)
+// BLSSignerFn is a signer callback function to request a message and extra data to be signed by a
+// backing account using BLS with a direct or composite hasher
+type BLSSignerFn func(accounts.Account, []byte, []byte, bool) (blscrypto.SerializedSignature, error)
 
 // Backend provides application specific functions for Istanbul core
 type Backend interface {
@@ -75,8 +71,9 @@ type Backend interface {
 
 	// Sign signs input data with the backend's private key
 	Sign([]byte) ([]byte, error)
-	SignBlockHeader([]byte) (blscrypto.SerializedSignature, error)
-	SignBLSWithCompositeHash([]byte) (blscrypto.SerializedSignature, error)
+
+	// Sign with the data with the BLS key, using either a direct or composite hasher
+	SignBLS([]byte, []byte, bool) (blscrypto.SerializedSignature, error)
 
 	// CheckSignature verifies the signature by checking if it's signed by
 	// the given validator
@@ -104,5 +101,5 @@ type Backend interface {
 	RefreshValPeers() error
 
 	// Authorize injects a private key into the consensus engine.
-	Authorize(address common.Address, publicKey *ecdsa.PublicKey, decryptFn DecryptFn, signFn SignerFn, signHashBLSFn BLSSignerFn, signMessageBLSFn BLSMessageSignerFn)
+	Authorize(address common.Address, publicKey *ecdsa.PublicKey, decryptFn DecryptFn, signFn SignerFn, signBLSFn BLSSignerFn)
 }
