+++ consensus/istanbul/backend/val_enodes_share_test.go
@@ -30,7 +30,7 @@ func TestHandleValEnodeShareMsg(t *testing.T) {
 		t.Errorf("error %v", err)
 	}
 
-	b.Authorize(getAddress(), &privateKey.PublicKey, decryptFn, signerFn, signerBLSHashFn, signerBLSMessageFn)
+	b.Authorize(getAddress(), &privateKey.PublicKey, decryptFn, signFn(nil), signBLSFn(nil))
 
 	// Set the backend's proxied validator address to itself
 	b.config.ProxiedValidatorAddress = senderAddress
