+++ consensus/istanbul/core/commit.go
@@ -32,7 +32,7 @@ func (c *core) sendCommit() {
 
 func (c *core) generateCommittedSeal(sub *istanbul.Subject) (blscrypto.SerializedSignature, error) {
 	seal := PrepareCommittedSeal(sub.Digest, sub.View.Round)
-	committedSeal, err := c.backend.SignBlockHeader(seal)
+	committedSeal, err := c.backend.SignBLS(seal, []byte{}, false)
 	if err != nil {
 		return blscrypto.SerializedSignature{}, err
 	}
@@ -78,7 +78,7 @@ func (c *core) broadcastCommit(sub *istanbul.Subject) {
 	}
 	var epochValidatorSetSeal blscrypto.SerializedSignature
 	if err == nil {
-		epochValidatorSetSeal, err = c.backend.SignBLSWithCompositeHash(epochValidatorSetData[:])
+		epochValidatorSetSeal, err = c.backend.SignBLS(epochValidatorSetData[:], []byte{}, true)
 		if err != nil {
 			logger.Error("Failed to sign epoch validator set seal", "err", err)
 			return
