+++ consensus/istanbul/core/testbackend_test.go
@@ -73,7 +73,7 @@ type testCommittedMsgs struct {
 //
 // define the functions that needs to be provided for Istanbul.
 
-func (self *testSystemBackend) Authorize(address common.Address, _ *ecdsa.PublicKey, _ istanbul.DecryptFn, _ istanbul.SignerFn, _ istanbul.BLSSignerFn, _ istanbul.BLSMessageSignerFn) {
+func (self *testSystemBackend) Authorize(address common.Address, _ *ecdsa.PublicKey, _ istanbul.DecryptFn, _ istanbul.SignerFn, _ istanbul.BLSSignerFn) {
 	self.address = address
 	self.engine.SetAddress(address)
 }
@@ -121,22 +121,11 @@ func (self *testSystemBackend) Multicast(validators []common.Address, message []
 	return nil
 }
 
-func (self *testSystemBackend) SignBlockHeader(data []byte) (blscrypto.SerializedSignature, error) {
+func (self *testSystemBackend) SignBLS(data []byte, extra []byte, useComposite bool) (blscrypto.SerializedSignature, error) {
 	privateKey, _ := bls.DeserializePrivateKey(self.blsKey)
 	defer privateKey.Destroy()
 
-	signature, _ := privateKey.SignMessage(data, []byte{}, false)
-	defer signature.Destroy()
-	signatureBytes, _ := signature.Serialize()
-
-	return blscrypto.SerializedSignatureFromBytes(signatureBytes)
-}
-
-func (self *testSystemBackend) SignBLSWithCompositeHash(data []byte) (blscrypto.SerializedSignature, error) {
-	privateKey, _ := bls.DeserializePrivateKey(self.blsKey)
-	defer privateKey.Destroy()
-
-	signature, _ := privateKey.SignMessage(data, []byte{}, true)
+	signature, _ := privateKey.SignMessage(data, extra, useComposite)
 	defer signature.Destroy()
 	signatureBytes, _ := signature.Serialize()
